Better errors and logging.
errors now tell you what we think it is if its something we know. also
added and option for logging the output into a log file. a small bug was
also fixed where sometimes the images have both the tag and digest so we
had to remove the tag to be able to pull using docker